Initial commit. First (working) version.
This commits the first working version of imgpacker. It is packing images.
The mark for this commit is that the algorithm is working for the initial
batch of tests I've done.

The program is still incomplete. It needs to output CSV files yet.

The program is already inconsistency on some minor points with the pre
analysis and implementation design (see files pre_analysis.txt and
pre_impl_design.txt for more details).